poor (hu)man's consensus for bucket types
Cluster Metadata's eventually consistency and bucket type semantics
don't play well together. This commit addresses this disharmony by
adding some hacky consensus. It does this in two steps. First, all
writes to bucket types are now serialized through the claimant. Second,
the status of the bucket type can be checked. See riak_core_bucket_type:status/1
for more details. Using the status, a user can decide whether its "safe" to
update the type (or determine if it has been successfully created on all nodes,etc).
Many of the cases where updating a type is not recommended in the docs of status/1
are safe to update only if the claimant remains alive (or goes down but comes back up
w/o the user downing the node). However, this is error prone and brittle,
so these cases are not recommended.

One case not handled by this commit is if a user does indeed go against recommendations
and downs a claimant when all recent bucket type's update have not reached the new claimant.
The creation semantics in this case would be undefined.